AU: Use full rootfs partition for scratch.
This patch allows the delta generator to use blocks beyond rootfs up to
the full rootfs partition as scratch. This is really a stop gap solution.
The updater needs to be fixed to work with as little as one block of
scratch if necessary.

BUG=6531
TEST=unit tests, generated a delta payload and updated from 70.8 to 72.3

Change-Id: I52b7d04d5a5345c34c9c647f9878c836be9f489c

Review URL: http://codereview.chromium.org/3562001