AU: Delta Performer: properly detect idempotent operations
Idempotent operations can be restarted after being aborted. This CL
enables the delta performer to find many idempotent operations. The
algorithm given is pretty good, tho it may incorrectly mark some
idempotent operations as non-idempotent. For example, a BSDIFF
operation may write a block that it reads with the same data; that's
idempotent, but missed by this algorithm. This function just looks to
see if the src/dst blocks are non-intersecting and, if non
intersecting, declares it idempotent.

BUG=7394
TEST=unittest

Review URL: http://codereview.chromium.org/3591019