AU: Really delta compress the kernel if an old kernel is provided.
Before this change, we always sent the full kernel (bzipped). Now,
if an old kernel is provided, we select the cheapest way to send
the kernel. If a full kernel update is really needed, the user
has the option of not providing an old kernel.

BUG=7705
TEST=unit tests, emerged on the device

Change-Id: I8dbfbfdf690fcbb50d5e6b5db52d994e448d5d25

Review URL: http://codereview.chromium.org/3767002