Verify downloaded size matches before verifying download hash match.
Hash mismatches should be rare and should not be seen during
normal operations, whereas size mismatches are more common due to failed
download attempts. So this CL re-orders the existing checks so that
we can easily determine the true number of hash mismatches, if any,
that's happening in the field.

BUG=chromium-os:30579
TEST=Tested with both valid and invalid sizes on my zgb.
Change-Id: Iea6b494011a32fbb3e30089eb7279db4927203d5
Reviewed-on: https://gerrit.chromium.org/gerrit/22067
Commit-Ready: Jay Srinivasan <jaysri@chromium.org>
Reviewed-by: Jay Srinivasan <jaysri@chromium.org>
Tested-by: Jay Srinivasan <jaysri@chromium.org>