Use IsDeltaPayload instead of IsDelta.
Trying to resurrect the IsDelta in an earlier CL to avoid pollution of
Omaha rules turned out to be a wrong choice because R14 and R15 code
strictly expects IsDelta to be true even for full payloads. So
preserving IsDelta's semantics as-is and introducing IsDeltaPayload to
really convey the intended semantics: True if it's delta payload, False
if it's full payload.

BUG=chromium-os:37801
TEST=Updated unit tests. Works fine on ZGB with Omaha devserver.

Change-Id: I899b56be4dcbdcf1981f72d356d40a1e99fdf725
Reviewed-on: https://gerrit.chromium.org/gerrit/41181
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Jay Srinivasan <jaysri@chromium.org>
Tested-by: Jay Srinivasan <jaysri@chromium.org>