feat(sys-kernel/coreos-bootkernel) Add a "bootloader" kernel package
This creates a sys-kernel/coreos-bootkernel package and eclass that are
really just a copy of the existing coreos-kernel package.  But, in the
future, this is going to start to diverge, both in the kernel
configuration, and the version of the kernel that is being used.

This also will provide a place to start building the needed initramfs
for this bootkernel that the "normal" kernel will not need.