fix(cros_sdk): Update default SDK tarball version.
Update to latest release. cros_sdk is still using this file instead of
version.txt in the manifest because it needs a tarball that already
exists when building a new version for the first time. At some point
that will work differently but for now we'll have to keep this file.