fix(app-editors/qemacs): Fix malformed patch.
This patch uses relative paths that are rejected by newer and stricter
eclass versions. Pulled from current portage but I'm purposefully *not*
updating the ebuild as the latest texi2html which in turn pulls in a lot
of new perl dependencies and we don't even care about texi2html or the
html docs it is used to produce in the first place so bleh.