fix(coreos-base/verity): Fix use of gcc flag not available in 4.6
Verity uses -fstack-protector-strong which was added in 4.7, use -all
instead which is available in 4.6. Dunno what impact this has but we
don't actually use it yet so *shrug*. Only a temporary fix since we
would probably want to fork verity if we use it or drop support entirely
if we decide we never will.