fix(app-emulation/virtualbox): Support minimal virtualbox tools builds
We don't actually want to run virtual machines, just get at the tools so
we can build virtualbox and vagrant images. So we disable nearly all use
flags and remove the virtualbox-modules dependency. For all versions
other than virtulbox-9999 the modules ebuild creates the vboxusers
group, so add that to the other virtualbox ebuilds instead.

Some of the shuffling in profiles is due to make.defaults appears to win
over package.use when they are in the same profile directory and we need
to disable the python flag because for some reason the virtualbox
configure script cannot find ours. To get the profile working right the
python flag is now turned on in base instead so the package.use in
default/linux can override it. I really look forward to cleaning up
these profiles once and for all, this is a mess.