add(sys-kernel/coreos-firmware): Add firmware package.
This provides firmware from linux-firmware but excludes everything not
explicitly required by coreos-kernel. Note that firmware installed by
this will only be available on the root filesystem, the initrd still
uses the smaller set of firmware the linux repo provides.

The current 3.12 kernel wants a few files missing in the July snapshot:

 * Missing firmware: ct2fw-3.2.1.1.bin (bna.ko)
 * Missing firmware: ctfw-3.2.1.1.bin (bna.ko)
 * Missing firmware: cxgb4/t5fw.bin (cxgb4.ko)

These files have been added to the linux-firmware repo so a newer
snapshot will be required but for now I'm sticking with what is already
in Gentoo so I don't have to generate my own tarball or whatever.