hack(sys-libs/e2fsprogs-libs): assert compile_et is being installed
A broken e2fsprogs-libs binary package lacking compile_et has been
causing problems. I am entirely at a loss as to how this happened in the
first place and unfortunately the error is not revealed until much
later. Hopefully this crude test at the end of src_install will find the
error as it happens. Or at the very least the revision bump will force
everything to move past the one bad build.