fix(sys-apps/systemd) allow booting without /etc/os-release
systemd will not boot if /etc/os-release is not present.

We "know" better, and can recover from a lost /etc/ directory, so patch
out the test.  This is only for "real" booting, not for nspawn
containers at this point in time as we aren't booting full system images
that way.  Yet.