fix(coreos-base/oem-vagrant): Split up plugin, use change_host_name hook.
The old override-plugin.rb didn't actually properly monkey patch the
change host name function because it was missing a requires line
pointing to the file it is defined in upstream. To clarify things split
the code into two files that match the upstream names. Additionally
using this hook as it is intended means it works if a host name is
defined but additional networks are not.

As part of the split switch to starting the units directly. The
indirection through a path unit just adds complexity.