remove(metrics): Remove calls to metrics library.
We do not use the ChromeOS metrics daemon to gather system information.
The limited info we do gather is just reported to the update service.
Note: this does not remove the code that generates the info being
submitted, just the calls themselves, the rest will go in a follow up.