fix(profiles): gut board package.provided
The intent of this file is to exclude dependencies from both the build
root in /build/amd64-usr and board images that would be pulled in but
aren't actually needed. Many of these are build utilities that need to
be provided by the SDK rather than the build root. The current state is
a bit confusing because we are actually including things like git but it
is also listed here and this somehow works due to version differences.

Additionally dev images will start including a working build environment
soon so may of these tools need to get built after all. To simplify
things everything has been removed except for the few packages that are
would get pulled into production images if it weren't for this file.

packages.provided actually shouldn't be used in this way, instead
controlling dependencies should be done via USE flags and possibly some
day the experimental HDEPEND ebuild variable.