payloads: remove original signing key version
OrignalVersion was only used to generate a warning when creating a
update payload and mis-used in a couple tests. The warning is just going
to be hidden under a couple layers of scripts (delta_generator's usage
is far too terrible for anyone to use it by hand) so the warning is
unlikely to serve any use and would just contribute noise while
generating updates with dev keys which only need a single version.