http_fetcher: remove all remaining proxy handling code
Instead of implementing a new ProxyResolver class that reads environment
variables we can support the feature simply by removing the mechanism
entirely. libcurl has built in support for reading proxy information
from the environment itself but was overridden in LibcurlHttpFetcher.

This does eliminate the mechanism for trying multiple proxies or and
attempting direct connections periodically to work around broken
proxies. Arguably CoreOS actually shouldn't try to do this. Unlike an
average ChromeOS user an admin that configures their infrastructure to
use a proxy is more likely to notice and get grumpy if CoreOS randomly
ignores the proxy and attempts to connect directly.