spidermonkey: fix cross-compiling logic
The existing ebuild uses a really crazy hack for cross compiling which
may have worked a few versions ago but it doesn't now. The root issue is
that Mozilla mixes up the meaning of "host" and "target" so give in to
their stupid and setup the environment with their meaning.

The configure script claimed in a comment:

   In Mozilla, we use the names $target, $host and $build incorrectly,
   but are too far gone to back out now. See Bug 475488:
     - When we say $target, we mean $host, that is, the system on which
       Mozilla will be run.
     - When we say $host, we mean $build, that is, the system on which
       Mozilla is built.
     - $target (in its correct usage) is for compilers who generate
       code for a different platform than $host, so it would not be used
       by Mozilla.

I'm inclined to smack someone with a stick.