omaha_request_action: do not generate useless core dumps
When HTTP reports success but the Omaha response includes an error just
log an error and move on instead of triggering a pointless core dump.

The weirdest thing is the core dump is made via a call back from the
main event loop so the dump doesn't actually include any context for
what triggered the dump. Gone now. :)