profiles: trim more config files from /etc
Some of these were deleted by build_image, others were still being
shipped but aren't really needed.

The big question mark is LVM, it isn't clear if LVM's default behavior
is actually sane or if the configs are needed to make it sane. Either
way we were already removing this, but something to note in case issues
crop up eventually.