bash: latest security fixes (work in progress)
This is the adds two more patches that were posted this morning. I
haven't seen anyone declaring that they are insufficient yet but they
may not be the final versions either.

OOB memory access:
http://www.openwall.com/lists/oss-security/2014/09/25/32

Export functions with an extra prefix:
http://www.openwall.com/lists/oss-security/2014/09/25/13

I'm guessing the first is safe, the later offers a more complete fix for
the previous two code execution issues but may cause problems for
portage. I'm testing that theory now.