profiles: do not install live dtc ebuild
Not a significant feature here, it is an OpenFirmware compiler and
optional dependency of QEMU. Old ChromeOS stuff was allowing the live
ebuild and the project's upstream host seems to now be gone, breaking
our builds. Remove keyword, remove from hard-host-depends. It can be
pulled in as a dependency of QEMU if needed.