sys-kernel/coreos-sources: fetch kernel sources via tarballs
Providing the kernel source via git is overkill, we do not maintain any
patches on top of the mainline kernel any more and the repository is
huge. As an added benefit this installs the kernel source to the
traditional /usr/src/linux location that Gentoo expects. Hopefully that
will reduce the number of warnings many ebuilds tend to output.