profiles: enable ipc and network sandboxing
Another day, another intermittent build failure because the local xml
docbook catalog didn't work right and the xml parser also failed to
fetch the schema over the network. Configure portage to use network
namespaces to prevent such situations from ever working in the first
place so at least the errors are unlikely to be intermittent.

For the sake of completeness use ipc namespaces too but that is less
likely to actually be a factor in our builds.