coreos-kernel: new kernel build based on coreos-sources
This cuts out the need for git for our kernel sources and the + and the
end of our versions is now gone. The build remains a bit weird in order
to bundle the initrd into the kernel image but eventually that will be
dropped, leaving the new coreos-kernel eclass pretty tiny.

Bonus: That + at the end of our versions is finally gone!