scripts: do not include board binaries in $PATH
The wrapper for `foo-config` style scripts updated `$PATH` so it could
call the real script by name instead of full path. However this means
that calls to shell utilities like `sed` may use the board's version
instead of the build host's. This hasn't been a problem in the past
because the board and host happened to be compatible but this may not
always be true.

In other news, come on people, switch to pkg-config already.