dev-libs/libdivsufsort: remove bogus mirror restriction
This was mistakenly added in the early turmoil of forking ChromiumOS in
commit 86ce9947: "the gentoo mirrors don't have this package, download
from Google code for now." Since then we have started to provide our own
mirror instead of using Gentoo's.

For reference, the better rule to use is `RESTRICT=primaryuri` which
prioritizes downloading from `SRC_URI` over the distfiles mirrors but
does not disallow uploading to a mirror.