profiles: block packages with distribution issues
Gentoo has switched to using USE=bindist only for packages with optional
features that cannot be distributed and RESTRICT=bindist otherwise. This
change doesn't impact anything in our tree at the moment but we should
follow along to avoid mistakes in the future.

Also, block packages with mirror restrictions. This is just to call
attention to anything that would otherwise be missing from our distfiles
mirror. Nothing we ship should actually have such a restriction.