connection_manager: remove all ConnectionManager related code
This mechanism was used to optionally disable updates of Chrome systems
connected over cellular networks. This isn't applicable to us and
touches a lot of code so time to clean it out.

After this HttpFetcher classes no longer need a reference to the global
SystemState so that plumbing has also been removed.

The external library, 'libchromeos' is no longer required.