omaha_request_action: remove support for deferring updates
To support ChromeOS users wanting to delay an update until later and
similar use cases there is a lot of goo to allow checking for an update
but do nothing for a limited period of time or a limited number of
update checks. Neither of these mechanisms are used in CoreOS and the
only related use case, only allowing updates during defined maintenance
windows, would be implemented much differently. Bye!