build: detect which namespace gflags uses
gflags 2.0 and earlier used the `google` namespace while versions 2.1
and 2.1.1 switch to the `gflags` namespace, breaking both API and ABI
compatibility. Then 2.1.2 restores 2.0 ABI compatibility and provides a
`GFLAGS_NAMESPACE` macro.  Considering a year passed between 2.1.1 and
2.1.2 and the latter still isn't marked stable on Gentoo we define the
macro if it isn't provided.

Use of pkg-config has been removed because 2.1 and later dropped
libgflags.pc in the process of moving to cmake. :(