strings: replace number conversions implementation with strtol
Although it would have been nice to swap these for the C++11 std::stoi
and related functions, the std functions throw exceptions which are
currently not supported by the update_engine code base. We don't need
most of the functionality of the old code so strip it out and replace
with strtol functions, adding extra error checking as appropriate.