eclass/coreos-go: Add cross build support
To support golang cross builds add a new function go_get_arch() for ebuilds
to use as needed.  Also update coreos-go_src_prepare() to set the go compiler
environment variable GOARCH to the result of go_get_arch().  The setting of
GOARCH will cause the compiler to generate the correct binary type for the
board.

Signed-off-by: Geoff Levand <geoff@infradead.org>