strings: add simplified TrimWhitespace function
The original version of this provided by Chromium's string_util accepted
an argument for trimming the beginning, end, or both ends of the string.
Additionally it reported whether any characters were actually trimmed
via a return value.  All calls in update_engine trim both ends and
ignore the return value so drop the extra arg and complication. Adding
this to string_split since it is the only piece of string_util we will
be importing and it is used there.