strings: unify character and substring splitting functions
SplitString and SplitStringUsingSubstr are very similar, one simply
being a special case where the delimiter substring is a single char.
The two implementations behave slightly differently, the Substr version
returns a vector containing a single empty string when given empty
input. That seems odd so move to the other behavior of returning an
empty vector. If there is any performance advantage to the single
character version it isn't really important for this code base.

While revising this code take the opportunity to modernize, returning
the vector directly instead of using an output parameter.