coreos-postinst: fix quoting in app id argument
When coreos-postinst is called by the current update_engine which
provides the app id as a second argument the value was getting an extra
'}' at the end due to bash just looking for the first '}' char.

Update the log message at the end to include app id and don't bother
reading the lsb-release file from the ancient amd64-generic images.