coreos-firmware: do not exclude directories and symlinks
Commit ed4660a2d9abe4583b9b327d2801703ec9f4ca18 added an explicit error
message to catch when something goes wrong and cannot find any firmware
required by the kernel. This additionally changed the call to doins to
an explicit list of files, flattening out the directory tree and
excluding symlinks. The tree *must* be preserved exactly in order for
the kernel to find the files it needs.