scoped_file: log close errors but do not abort process
update_engine does not need ScopedFD's "abort on failure" behavior, the
previous implementation from utils.h doesn't even log errors. Somewhere
in between the two implementations is the more reasonable behavior of
just logging the error and moving on.

For extra fun ScopedFILE did not share this behavior and ignored errors.
Sync it up with ScopedFD and add a unit test.