delta_performer: always use operation hash checks
The "checks mandatory" toggle only really made sense for the manifest
hash and signature in Omaha responses. However we removed this
functionality to make key rotation more practical in de796c7b. The
operation hashes contained in the manifest are always there and don't
depend on anything special from Omaha so allowing them to fail is silly.