full_update_generator: refactor to support kernels and other objects
Instead of adding new objects by expanding the argument list to a single
function as was done in the original kernel implementation, it'll be
much easier to read if each object is added to the update payload via
individual function calls. The existing all-in-one function remains for
now, code using it will be fixed/adapted later.

The test code is partially resurrected from the previous implementation.