coreos-postinst: search for ld.so instead of assuming amd64
Side note: given that postinst is shipped in the image along side cgpt
and ld.so we could define both of these paths at compile time instead of
searching at run time. Plumbing that through the ebuild, configure, and
this script is more than I want to deal with at just this moment though.