update_metadata: change procedure type field to optional
C++ protobuf treats unknown enum values and unknown fields. Thus if the
field is required parsing aborts with a missing required field error.
Since I want to be able to add new procedure types in the future and
have update_engine gracefully skip them we need to switch to optional.

Normally switching from required to optional isn't recommended but we
haven't released this version of update_engine for amd64 yet.