test_utils: use ext2 instead of ext3
Images should never be generated from a journaled filesystem. Attempting
to shove the 4MB journal in the test filesystem normally isn't a big
deal but sometimes it appears to tickle something problematic in the
bsdiff algorithm, taking as long as 15-20 minutes to run.

So lets just not do that shall we? :-P