delta_diff_generator: do not use area after filesystem as scratch space
To avoid having to plumb verity parameters through the update process in
order to regenerate the hashes we are just going to send the hashes in
the update payload itself. Since the update will then cover the entire
device there isn't any extra space to toss in as scratch.

This only changes the generation of delta updates which we currently
don't use. Additionally it shouldn't adversely impact deltas since there
are other ways for the generator to find scratch space.

Bonus: We no longer hard-code the partition size!!