kernel_verifier_action: allow payloads without kernels
The code partially supports gracefully skipping the kernel update if it
isn't in the payload but the verify sequence didn't. We could
potentially change this later if everything that can generate payloads
starts to always provide the kernel.

Log the kernel size and hash when successful like the filesystem
copier/verifier does.