coreos-base/oem-gce: change some host bind mounts
This drops two mounts that are not needed anymore, since the ACI is
built with emerge and gets those data dependencies now.  It adds a
new mount for home directories to be created.

Two existing mounts remain: one for /etc, and one for /run/systemd
so the /dev/log link works.