omaha_request_action: use manifest version attribute
CoreUpdate does not send a "DisplayVersion" attribute on the postinst
node, which is not standard. However it does send the standard "version"
attribute on the manifest node so we should just use that instead.

Deprecate and remove support for the non-standard DisplayVersion.
Internally the field can remain named "display_version" since it is
purely informational.