coreos-postinst: patch platform.py in python libs
Azure's agent erroneously looks at the distribution name instead of its
ID. In order to rename the OS to "Container Linux", this fix is needed
to force python to always report the OS's name as "CoreOS".

This will apply to all of the OEMs where we ship python (historically,
Azure and GCE), but Azure is the only one who uses these APIs.
Eventually, the Azure agent will be updated to read the ID instead of
the name, making this unnecessary on newer instances. Unfortunately,
since we don't update the OEM partition, there will always be older
agents who require this fix.