app-emulation/runc: re-apply non-dumpable patch
This re-fixes CVE-2016-9962

Due to reusing the same "${P}" as upstream gentoo and also using their
distfile mirror, the downloaded source tarball was not for the
EGIT_COMMIT referenced in the ebuild, and so did not include the patch.

I've also updated the EGIT_COMMIT to match what's really being used,
though it's still ignored in this change.

The patchfile was taken from gentoo's overlay.

This was done rather than fixing the ebuild to use the intended commit
simply because it ends up being a smaller change. The proper change to
use the commit intended will be done on other channels.