sys-fs/lvm2: apply CoreOS changes and bump
There is a dependency loop when systemd is build with cryptsetup
like systemd -> cryptsetup -> lvm2 -> baselayout -> systemd.  Since
CoreOS always installs a good baselayout, it makes sense drop the
lvm2 dependency and turn it into a conflict with bad versions.  It
also has a runtime dependency loop on systemd, so make that into a
build dependency.

This also stabilizes the version that was being used by arm64,
updating amd64, to bring the architectures into alignment.