sys-process/tini: add tini ebuild
Even though there's an upstream gentoo ebuild, it depends on cmake
(which had trouble cross-compiling for arm64).

tini is a simple enough program that writing our own makefile for it is
pretty easy, and it'll pay off in reduced buildtime and dependencies for
now since we can avoid dragging in cmake.

Note that we don't bother providing the static useflag and just make it
always static since we're packaging this for docker regardless.