app-emulation/docker: Sync with upstream, bump to 17.06.1 / Go 1.8
This switches to the new Docker CE repo structure.  It points to
the docker/docker-ce GitHub repo directly now, not the CoreOS fork.

The cros_workon code was dropped, so we are building unmodified
upstream releases now.  If fixes need to be backported, patch files
can just be added to the PATCHES array since this uses EAPI=6.

LDFLAGS are tacked onto the -extldflags option to support torcx.  A
special RPATH is added for torcx support, which is passed through
the LDFLAGS environment variable.  The value is applied along with
the "hardened" USE flag's options, so note that torcx will not work
if that flag is disabled.