service: drop file
We only listen on the system bus, and the User key is mandatory there.
Thus this file has never been relevant. It's too late to add User and
SystemdService keys and start using proper bus activation, because
we document that stopping update-engine.service will prevent updates
from occurring until the machine is rebooted.