coreos-base/update_engine: drop D-Bus service file
We've always installed it in the session bus directory, which is
useless, since update_engine only communicates on the system bus.  This
prevented bus activation from working. We can't fix bus activation now
because that could cause update_engine to start in cases where the user
believes it to be disabled, so just drop the file.