coreos-base/update-ssh-keys: add update-ssh-keys
this adds the update-ssh-keys package, which is a utility for managing
groups of openssh authorized public keys. the previous incarnation of
this utility included in the operating system was a shell script
included in init. it is now a rust binary living at
github.com/coreos/update-ssh-keys.

The update-ssh-keys shell script is removed from init in
ed478168f0133cc140baa08d4d77e4333e8991d7, github.com/coreos/init#256.