sys-apps/systemd: "Fix" telinit in the SDK
When telinit is run and doesn't detect the host systemd, it falls
back to executing a default fallback program that is detected by
meson.  In this case, it detects itself at /sbin/telinit, so it
enters an infinite exec loop.  Change the default fallback program
to a non-existent path so that telinit just fails in the SDK.