profiles: Disable ccache for SDK packages
This profile is used by catalyst when building toolchains packages,
which randomly fail from ccache permissions problems after certain
releases.

An alternative solution is to create a separate profile just for
catalyst, but that doesn't seem necessary given that there aren't
many SDK packages that would benefit from this caching that aren't
able to be tested on the amd64-usr board, which still has ccache.