sys-libs/glibc: Work around usrmerge problems
Portage will blindly clobber like-named files when extracting /lib
and /usr/lib if they are symlinked together on the host.  In this
case, it extracts a symlink that points to itself, so this changes
it to move the library into /usr for the boards.  Checks in portage
and the ChromeOS scripts prevent any straightforward copying or
linking.