app-admin/sudo: Apply CoreOS changes
Stabilize, adjust LDAP support, and add SSSD sudoers support.

This drops the Perl dependency for LDAP since it seems like it was
only pulled in for an optional script that is no longer present.
Also, the schema files are dropped from the installation since our
OpenLDAP package has USE=minimal which skips the schema directory.
(It still installs a default config file in /etc, but it contains
only comments since there are a few others like that already.)

Gentoo is leaving the SSSD option stalled in #525674 since 2014, so
maybe this can be moved back to portage-stable if that is merged.
Note that the proposed change there will depend on SSSD, so it will
have to be disabled in the arm64 profile in that case, unless the
SSSD packages are fixed by then.