sys-auth/google-oslogin: add package
Add a package for google cloud's oslogin functionality. This adds the
pam and nss modules but not any configuration to activate them. That
configuration will live in the coreos-base/oem-gce package.

Furthermore, despite being gce specific, this will be included in USR
since it is security critical and may need to be updated.