sys-libs/glibc: Apply CoreOS changes
Drop pkg_pretend since it breaks build_image if cross-compilers are
not installed yet (e.g. in Jenkins jobs).

Drop the libidn2 runtime dependency since it breaks bootstrapping,
and it's dlopen()ed so the resolver can work without it.

Drop the host /dev/pts checks since the SDK doesn't control it.

Apply our gshadow segfault patch.  (We no longer steal Fedora's
patch for C UTF-8 since it's in the Gentoo patch set now.)

Install nscd.conf in /usr and set up tmpfiles to link it in /etc.

Wipe out /etc files (except for an environment file that is still
needed in the SDK).